+++ src/Blammo/Logging/WithLogger.hs
@@ -1,4 +1,4 @@
-module Blammo.Logging.WithLogger (WithLogger (..)) where
+module Blammo.Logging.WithLogger (WithLogger (..), runWithLogger) where
 
 import Prelude
 
@@ -12,6 +12,9 @@ import Control.Monad.Reader (MonadReader, ReaderT (ReaderT), asks)
 --   'MonadLogger' for your application monad
 newtype WithLogger env m a = WithLogger (ReaderT env m a)
   deriving newtype (Functor, Applicative, Monad, MonadIO, MonadReader env)
+
+runWithLogger :: env -> WithLogger env m a -> m a
+runWithLogger env (WithLogger (ReaderT f)) = f env
 
 instance (MonadIO m, HasLogger env) => MonadLogger (WithLogger env m) where
   monadLoggerLog loc source level msg = do