+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [v1.1.4.0](https://github.com/freckle/blammo/compare/v1.1.3.0...v1.1.4.0)
 
-- New in `Blammo.Logging`: `withLogger`, `WithLogger(..)`
+- New in `Blammo.Logging`: `withLogger`, `WithLogger(..), runWithLogger`
 - New in `Blammo.Logging.Logger`: `runLogAction`
 
 ## [v1.1.3.0](https://github.com/freckle/blammo/compare/v1.1.2.3...v1.1.3.0)